requirements.txt
datasets==2.20.0
deepspeed==0.15.1
dill==0.3.8
evaluate==0.4.3
# faiss_gpu==1.7.2
filelock==3.16.1
hydra-core==1.3.2
inflection==0.5.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
peft==0.14.0
Pillow==10.4.0
pycocoevalcap==1.2
pycocotools==2.0.8
pytest==8.3.3
pytorch_lightning==2.4.0
rich>=10.2.2
setuptools==69.5.1
termcolor==2.5.0
torch==2.5.1
tqdm==4.66.5
transformers
wandb>=0.12.10
xxhash==3.4.1
